GameCenter fix 4th attempt
instead of using (Transform highscoreText) to take the score value from I'm using Score.scoreValue from Score class. I changed public static int scoreValue = 0; to long to fix error I was getting

+++ Assets/Score.cs
@@ -5,7 +5,7 @@ using UnityEngine.UI;
 
 public class Score : MonoBehaviour
 {
-    public static int scoreValue = 0;
+    public static long scoreValue = 0;
     Text score;
     // Start is called before the first frame update
     void Start()
